Fixed day/night times for economy 7 tariffs when using a smart meter.
According to the the page below the off peak night rate is at a different
time if you have a smart meter. This patch adds a config option to select
if a smart meter is installed, and uses that to calculate the correct
time to change to the night rate.
https://octopus.energy/help-and-faqs/articles/what-happens-to-my-economy-seven-e7-tariff-when-i-have-a-smart-meter-installed/