feat: updated the default names of various sensors to be more useful in areas of HA when space is limited
BREAKING CHANGE:
If you are relying on the previous default names of sensors then you will need to update your
sensors manually. Only the visual default name of the sensors have changed; the entity ids will
remain the same and therefore will not effect any automatmations. If you have manually changed the
names of the sensors, this should not effect you.