Use MPRN to get Gas override tariff
If only gas is supplied by Octopus, MPAN is not defined,
so the call fails with the error:
`cannot access local variable 'mpan' where it is not associated with a value`

By the looks of it, MPRN should be used instead.

I tried it on my local setup and the gas-only integration works fine.