feat: updated intelligent target time (time based) sensor to be disabled by default
BREAKING CHANGE:
The intelligent target time (time based) sensor is disabled by default. If you are using this
variation of the sensor, you might need to re-enable the sensor and you should disable the select
variation. If you are using the select variation, the time variation should be disabled. This is
instead of removing the time variation as previously advertised. See docs for side effects of these
sensors.