Update Services Documentation to avoid API call clumping
The example provided will clump requests at the 3 minute mark, assuming they follow the example without considering the impact this may have on the Octopus APIs. 
Automatic mode wouldn't clump, as it's tied to the startup time of home assistant, which provides a decent random offset in most scenarios. 

I'd suggest that users be encouraged to pick a random offset value for periodic requests, so that this doesn't cause a problem with the IO API getting overloaded